% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_inserts.R
\name{make_inserts}
\alias{make_inserts}
\title{Make SQL inserts from data.frame}
\usage{
make_inserts(data, table_name = NULL)
}
\arguments{
\item{data}{\code{data.frame} used to generate SQL inserts}

\item{table_name}{target table.}
}
\description{
Function produces sql inserts from given dataset
}
\examples{
make_inserts(mtcars)
}

