% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlinserts.R
\docType{package}
\name{sqlinserts-package}
\alias{sqlinserts}
\alias{sqlinserts-package}
\title{Produces SQL inserts from given dataset}
\description{
Produces SQL inserts from given dataset. Efficient implementation in C++.
}
\details{
Package helps to roduces SQL inserts from given \code{data.frame} object.
}

